Move store creation out of drafts to own file
