+++ src/store.ts
@@ -1,4 +1,7 @@
 
+import * as config from 'config'
+import {logger} from './logger'
+
 /**
  * Abstract blob store interface.
  * See https://github.com/maxogden/abstract-blob-store
@@ -83,4 +86,24 @@ export class AsyncBlobStore {
         })
     }
 
+}
+
+/** Main application store */
+export let store: AsyncBlobStore
+
+const storageConf: any = config.get('storage')
+if (storageConf.type === 'memory') {
+    logger.warn('using memory store')
+    store = new AsyncBlobStore(require('abstract-blob-store')())
+} else if (storageConf.type === 's3') {
+    const aws = require('aws-sdk')
+    const S3BlobStore = require('s3-blob-store')
+    const client = new aws.S3({
+        accessKeyId: storageConf.get('s3_access_key'),
+        secretAccessKey: storageConf.get('s3_secret_key'),
+    })
+    const bucket = storageConf.get('s3_bucket')
+    store = new AsyncBlobStore(new S3BlobStore({client, bucket}))
+} else {
+    throw new Error(`Invalid storage type: ${ storageConf.type }`)
 }
